finished exercises 4.15 - 4.23

+++ part4/bloglist/tests/blog_api.test.js
@@ -3,9 +3,14 @@ const supertest = require('supertest')
 const app = require('../app')
 const api = supertest(app)
 const Blog = require('../models/blog')
+const User = require('../models/user')
 const helper = require('./test_helper')
 
+let token
 beforeEach(async () => {
+    await api.post('/api/users').send({ username: 'damian', name: 'Damian', password: 'damian123' })
+    let login = await api.post('/api/login').send({ username: 'damian', password: 'damian123'})
+    token = login.body.token
     await Blog.deleteMany({})
     for (const blog of helper.initialBlogs) {
         await new Blog(blog).save()
@@ -31,10 +36,12 @@ test('blogs are successfully posted', async () => {
         likes: 50,
         url: 'http://blogpost.com/blogs_12_superman_is_best',
     }
-    const savedBlog = await api.post('/api/blogs').send(additionalBlog)
+    const savedBlog = await api.post('/api/blogs').set('Authorization', `Bearer ${token}`).send(additionalBlog)
 
-    const allBlogsResponse = await api.get('/api/blogs')
-    expect(allBlogsResponse.body).toContainEqual(savedBlog.body)
+    const allBlogs = await api.get('/api/blogs')
+    const addedPost = allBlogs.body.find(blog => savedBlog.body._id === blog._id)
+    
+    expect(allBlogs.body).toContainEqual(addedPost)
 }, 10000)
 
 test('blogs posted with missing likes property are assigned 0', async () => {
@@ -43,7 +50,7 @@ test('blogs posted with missing likes property are assigned 0', async () => {
         author: 'A. Qua',
         url: 'http://blogpost.com/blogs_12_aquaman',
     }
-    const savedBlog = await api.post('/api/blogs').send(additionalBlog)
+    const savedBlog = await api.post('/api/blogs').set('Authorization', `Bearer ${token}`).send(additionalBlog)
 
     const allBlogsResponse = await api.get('/api/blogs')
     const addedBlog = allBlogsResponse.body.find(blog => blog._id === savedBlog.body._id)
@@ -55,17 +62,17 @@ test('blogs posted with missing title and url properties are thrown 400 error', 
         author: 'Missing',
         likes: 1,
     }
-    await api.post('/api/blogs').send(additionalBlog).expect(400)
+    await api.post('/api/blogs').set('Authorization', `Bearer ${token}`).send(additionalBlog).expect(400)
 })
 
 test('blogs are successfully deleted', async () => {
     const allBlogs = await api.get('/api/blogs')
-    await api.delete(`/api/blogs/:${allBlogs.body[0]._id}`).expect(204)
+    await api.delete(`/api/blogs/${allBlogs.body[0]._id}`).set('Authorization', `Bearer ${token}`).expect(204)
 }, 10000)
 
 test('blogs are successfully updated', async () => {
     let allBlogs = await api.get('/api/blogs')
-    await api.put(`/api/blogs/:${allBlogs.body[0]._id}`).send({ likes: 100 })
+    await api.put(`/api/blogs/${allBlogs.body[0]._id}`).set('Authorization', `Bearer ${token}`).send({ likes: 100 })
     allBlogs = await api.get('/api/blogs')
     expect(allBlogs.body[0].likes).toBe(100)
 }, 10000)
